test: Integration test WIP
