--- src/ServiceLayer.Mesh/Functions/HealthCheck.cs
@@ -1,14 +1,0 @@
-using Microsoft.Azure.Functions.Worker;
-using Microsoft.Azure.Functions.Worker.Http;
-using Microsoft.Extensions.Logging;
-using System.Net;
-
-public class HealthCheckFunction
-{
-    [Function("HealthCheck")]
-    public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "health")] HttpRequestData req)
-    {
-        var response = req.CreateResponse(HttpStatusCode.OK);
-        return response;
-    }
-}
